fix: 新增购买接口

+++ bin-api-common/src/main/java/com/binbin/binapicommon/mode/dto/payInterfaceInfo/PayRequest.java
@@ -16,7 +16,7 @@ import java.io.Serializable;
  */
 @Data
 public class PayRequest implements Serializable {
-    private Integer interfaceInfoId;
+    private Long interfaceInfoId;
     private Integer payNum;
     private static final long serialVersionUID = -5842362260147083227L;
 }
